third_party/blink/web_tests/fast/filesystem/workers/simple-temporary-expected.txt

[Worker] requestFileSystem TEMPORARY test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/simple-temporary.js
[Worker] Successfully obtained TEMPORARY FileSystem:file__0:Temporary
PASS [Worker] fileSystem.name.length > 0 is true
PASS [Worker] fileSystem.root.fullPath is "/"
PASS successfullyParsed is true

TEST COMPLETE
